requirements.txt
SpeechRecognition 
PyAudio
google-generativeai
python-dotenv
flask
flask_cors
pymongo
